make lint

+++ worker/utils.go
@@ -26,7 +26,7 @@ func sumGasTxs(txs types.Txs) uint64 {
 	var totalGas uint64
 
 	for _, tx := range txs {
-		totalGas += uint64(tx.GasLimit)
+		totalGas += tx.GasLimit
 	}
 
 	return totalGas
